requirements.txt
python-telegram-bot
requests
pymongo[srv]
dnspython
beautifulsoup4
